+++ app/map/page.tsx
@@ -1,4 +1,4 @@
-import { MapView } from "./components/Map";
+import { MapView } from "./components/MapView";
 
 export default function MapPage() {
   return <MapView />;
